+++ card_live_dashboard/callbacks/__init__.py
@@ -2,6 +2,8 @@ from typing import List, Set, Dict
 from datetime import datetime, timedelta
 from dash.dependencies import Input, Output, State
 import dash
+from dash.exceptions import PreventUpdate
+import re
 
 from card_live_dashboard.service.CardLiveDataManager import CardLiveDataManager
 from card_live_dashboard.model.CardLiveData import CardLiveData
@@ -12,6 +14,8 @@ from card_live_dashboard.service.TaxonomicParser import TaxonomicParser
 DAY = timedelta(days=1)
 WEEK = timedelta(days=7)
 MONTH = timedelta(days=31)
+THREE_MONTHS = timedelta(days=365/4) # 3 months is defined as a quarter year
+SIX_MONTHS = timedelta(days=365/2) # 6 months is defined as half a year
 YEAR = timedelta(days=365)
 
 
@@ -53,9 +57,18 @@ def build_callbacks(app: dash.dash.Dash) -> None:
         return is_open
 
     @app.callback(
+        Output('custom-time-period', 'is_open'),
+        [Input('time-period-items', 'value')]
+    )
+    def toggle_custom_time_period(time_period_items):
+        return time_period_items == 'custom'
+
+    @app.callback(
         [Output('global-sample-count', 'children'),
          Output('global-most-recent', 'children'),
          Output('time-period-items', 'options'),
+         Output('date-picker-range', 'min_date_allowed'),
+         Output('date-picker-range', 'max_date_allowed'),
          Output('organism-lmat-select', 'options'),
          Output('organism-rgi-kmer-select', 'options'),
          Output('selected-samples-count', 'children'),
@@ -75,6 +88,8 @@ def build_callbacks(app: dash.dash.Dash) -> None:
          Input('organism-lmat-select', 'value'),
          Input('organism-rgi-kmer-select', 'value'),
          Input('time-period-items', 'value'),
+         Input('date-picker-range', 'start_date'),
+         Input('date-picker-range', 'end_date'),
          Input('timeline-type-select', 'value'),
          Input('timeline-color-select', 'value'),
          Input('totals-type-select', 'value'),
@@ -86,6 +101,7 @@ def build_callbacks(app: dash.dash.Dash) -> None:
                            amr_gene_families: List[str], resistance_mechanisms: List[str],
                            amr_genes: List[str], organism_lmat: str,
                            organism_rgi_kmer: str, time_dropdown: str,
+                           start_date: str, end_date: str,
                            timeline_type_select: str, timeline_color_select: str,
                            totals_type_select: str, totals_color_select: str,
                            rgi_type_select: str, n_intervals):
@@ -103,6 +119,21 @@ def build_callbacks(app: dash.dash.Dash) -> None:
         global_samples_count = len(data)
         global_last_updated = f'{data.latest_update(): %b %d, %Y}'
 
+        min_date_allowed = data.first_update()
+        max_date_allowed = datetime.now()
+
+        custom_date = {
+            'start': min_date_allowed,
+            'end': max_date_allowed
+        }
+
+        if start_date is not None:
+            start_date = datetime.strptime(re.split(r'[T ]', start_date)[0], '%Y-%m-%d')
+            custom_date['start'] = start_date
+        if end_date is not None:
+            end_date = datetime.strptime(re.split(r'[T ]', end_date)[0], '%Y-%m-%d')
+            custom_date['end'] = end_date
+
         time_subsets = apply_filters(data=data,
                                      rgi_cutoff_select=rgi_cutoff_select,
                                      drug_classes=drug_classes,
@@ -110,7 +141,8 @@ def build_callbacks(app: dash.dash.Dash) -> None:
                                      resistance_mechanisms=resistance_mechanisms,
                                      amr_genes=amr_genes,
                                      organism_lmat=organism_lmat,
-                                     organism_rgi_kmer=organism_rgi_kmer)
+                                     organism_rgi_kmer=organism_rgi_kmer,
+                                     custom_date=custom_date)
 
         fig_settings = {
             'timeline': {'type': timeline_type_select, 'color': timeline_color_select},
@@ -123,11 +155,12 @@ def build_callbacks(app: dash.dash.Dash) -> None:
         # Set time dropdown text to include count of samples in particular time period
         # Should produce a list of dictionaries like [{'label': 'All (500)', 'value': 'all'}, ...]
         time_dropdown_text = [{'label': f'All ({time_subsets["all"].samples_count()})', 'value': 'all'}]
-        for value in ['day', 'week', 'month', 'year']:
+        for value in ['day', 'week', 'month', '3 months', '6 months', 'year']:
             time_dropdown_text.append({
                 'label': f'Last {value} ({time_subsets[value].samples_count()})',
                 'value': value,
             })
+        time_dropdown_text.append({'label': 'Custom', 'value': 'custom'})
 
         samples_count_string = f'{time_subsets[time_dropdown].samples_count()}/{global_samples_count}'
 
@@ -147,6 +180,8 @@ def build_callbacks(app: dash.dash.Dash) -> None:
         return (global_samples_count,
                 global_last_updated,
                 time_dropdown_text,
+                min_date_allowed,
+                max_date_allowed,
                 organism_lmat_options,
                 organism_rgi_kmer_options,
                 samples_count_string,
@@ -163,7 +198,7 @@ def build_callbacks(app: dash.dash.Dash) -> None:
 def apply_filters(data: CardLiveData, rgi_cutoff_select: str,
                   drug_classes: List[str], amr_gene_families: List[str],
                   resistance_mechanisms: List[str], amr_genes: List[str], organism_lmat: str,
-                  organism_rgi_kmer: str) -> Dict[str, CardLiveData]:
+                  organism_rgi_kmer: str, custom_date: Dict[str, datetime]) -> Dict[str, CardLiveData]:
     time_now = datetime.now()
 
     data = data.select(table='rgi', by='cutoff', type='row', level=rgi_cutoff_select) \
@@ -179,8 +214,15 @@ def apply_filters(data: CardLiveData, rgi_cutoff_select: str,
         'day': data.select(table='main', by='time', start=time_now - DAY, end=time_now),
         'week': data.select(table='main', by='time', start=time_now - WEEK, end=time_now),
         'month': data.select(table='main', by='time', start=time_now - MONTH, end=time_now),
+        '3 months': data.select(table='main', by='time', start=time_now - THREE_MONTHS, end=time_now),
+        '6 months': data.select(table='main', by='time', start=time_now - SIX_MONTHS, end=time_now),
         'year': data.select(table='main', by='time', start=time_now - YEAR, end=time_now),
     }
+
+    if custom_date is not None:
+        time_subsets['custom'] = data.select(table='main', by='time', start=custom_date['start'], end=custom_date['end'])
+    else:
+        time_subsets['custom'] = data
 
     return time_subsets